Make all entries double

+++ diff.h
@@ -1,8 +1,8 @@
 #ifndef DIFF_H
 #define DIFF_H
 
-int ** calculateDifference(int *entries, int elength);
+double ** calculateDifference(double *entries, int elength);
 
-void printTable(int *entries, int elength, int ** table, int type); // type : 0 => forward difference, 1 => backward difference
+void printTable(double *entries, int elength, double ** table, int type); // type : 0 => forward difference, 1 => backward difference
 
 #endif
